+++ R/plot_comparisons.R
@@ -12,7 +12,7 @@
 #' * https://vincentarelbundock.github.io/marginaleffects/articles/plot.html
 #' * https://vincentarelbundock.github.io/marginaleffects
 #' 
-#' @param effect Name of the variable whose contrast we want to plot on the y-axis. If `NULL`, a plot of average comparisons is returned.
+#' @param variables Name of the variable whose contrast we want to plot on the y-axis.
 #' @param draw `TRUE` returns a `ggplot2` plot. `FALSE` returns a `data.frame` of the underlying data.
 #' @inheritParams comparisons
 #' @inheritParams plot_slopes
@@ -22,17 +22,17 @@
 #' @examples
 #' mod <- lm(mpg ~ hp * drat * factor(am), data = mtcars)
 #' 
-#' plot_comparisons(mod, effect = "hp", condition = "drat")
+#' plot_comparisons(mod, variables = "hp", condition = "drat")
 #'
-#' plot_comparisons(mod, effect = "hp", condition = c("drat", "am"))
+#' plot_comparisons(mod, variables = "hp", condition = c("drat", "am"))
 #' 
-#' plot_comparisons(mod, effect = "hp", condition = list("am", "drat" = 3:5))
+#' plot_comparisons(mod, variables = "hp", condition = list("am", "drat" = 3:5))
 #' 
-#' plot_comparisons(mod, effect = "am", condition = list("hp", "drat" = range))
+#' plot_comparisons(mod, variables = "am", condition = list("hp", "drat" = range))
 #' 
-#' plot_comparisons(mod, effect = "am", condition = list("hp", "drat" = "threenum"))
+#' plot_comparisons(mod, variables = "am", condition = list("hp", "drat" = "threenum"))
 plot_comparisons <- function(model,
-                             effect = NULL,
+                             variables = NULL,
                              condition = NULL,
                              by = NULL,
                              type = "response",
@@ -43,11 +43,22 @@ plot_comparisons <- function(model,
                              draw = TRUE,
                              ...) {
 
+
+    dots <- list(...)
+    if ("effect" %in% names(dots)) {
+        if (is.null(variables)) {
+            variables <- dots[["effect"]]
+        } else {
+            insight::format_error("The `effect` argument has been renamed to `variables`.")
+        }
+    }
     
     # sanity check
     checkmate::assert(
-        checkmate::check_string(effect),
-        checkmate::check_list(effect, names = "unique", len = 1))
+        checkmate::check_character(variables, names = "unnamed"),
+        checkmate::check_list(variables, names = "unique"),
+        .var.name = "variables")
+
     checkmate::assert_character(by, null.ok = TRUE, max.len = 3, min.len = 1, names = "unnamed")
     if ((!is.null(condition) && !is.null(by)) || (is.null(condition) && is.null(by))) {
         msg <- "One of the `condition` and `by` arguments must be supplied, but not both."
@@ -57,7 +68,7 @@ plot_comparisons <- function(model,
     # conditional
     if (!is.null(condition)) {
         modeldata <- get_modeldata(model, additional_variables = names(condition$condition))
-        condition <- sanitize_condition(model, condition, effect, modeldata = modeldata)
+        condition <- sanitize_condition(model, condition, variables, modeldata = modeldata)
         v_x <- condition$condition1
         v_color <- condition$condition2
         v_facet <- condition$condition3
@@ -69,7 +80,7 @@ plot_comparisons <- function(model,
             conf_level = conf_level,
             by = FALSE,
             wts = NULL,
-            variables = effect,
+            variables = variables,
             transform_pre = transform_pre,
             transform_post = transform_post,
             cross = FALSE,
@@ -86,7 +97,7 @@ plot_comparisons <- function(model,
             type = type,
             vcov = vcov,
             conf_level = conf_level,
-            variables = effect,
+            variables = variables,
             wts = NULL,
             transform_pre = transform_pre,
             transform_post = transform_post,
@@ -110,36 +121,7 @@ plot_comparisons <- function(model,
     # ggplot2
     insight::check_if_installed("ggplot2")
     p <- plot_build(datplot, v_x = v_x, v_color = v_color, v_facet = v_facet)
-
-    if (is.null(names(effect))) {
-        p <- p + ggplot2::labs(
-            x = v_x,
-            y = sprintf("Contrast in %s on %s", effect, {{condition$respname}}))
-    } else {
-        p <- p + ggplot2::labs(
-            x = v_x,
-            y = sprintf("Contrast in %s on %s", names(effect), {{condition$respname}}))
-    }
-
-    # `effect` is a categorical variable. We plot them in different facets
-    contrast_cols <- grep("^contrast$|^contrast_", colnames(datplot), value = TRUE)
-    for (con in contrast_cols) {
-        if (length(unique(datplot[[con]])) == 1) {
-            contrast_cols <- setdiff(contrast_cols, con)
-        }
-    }
-    if (length(contrast_cols) > 0) {
-        if (is.null(v_facet)) {
-            fo <- sprintf("~ %s", paste(contrast_cols, collapse = "+"))
-            p <- p + ggplot2::facet_wrap(fo)
-        } else {
-            fo <- sprintf("v_facet ~ %s", paste(contrast_cols, collapse = "+"))
-            p <- p + ggplot2::facet_grid(fo)
-        }
-    } else if (!is.null(v_facet)) {
-        fo <- stats::as.formula(paste("~", v_facet))
-        p <- p + ggplot2::facet_wrap(fo)
-    }
+    p <- p + ggplot2::labs(x = v_x, y = sprintf("Comparison"))
 
     return(p)
 }